Move user org unit type
