+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/analytics/common/dimension/DimensionParam.java
@@ -110,11 +110,6 @@ public class DimensionParam
         return Objects.nonNull( dimensionalObject );
     }
 
-    private boolean isQueryItem()
-    {
-        return !isDimensionalObject();
-    }
-
     /**
      * @return the DimensionParamObjectType of this DimensionParam
      */
