+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/common/AggregateAnalyticsQueryCriteria.java
@@ -35,7 +35,6 @@ import lombok.NoArgsConstructor;
 
 import org.hisp.dhis.analytics.AggregationType;
 import org.hisp.dhis.analytics.SortOrder;
-import org.hisp.dhis.analytics.UserOrgUnitType;
 
 /**
  * This class contains all the criteria that can be used to execute a DHIS2
