+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/common/UserOrgUnitType.java
@@ -25,9 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.analytics;
-
-import org.hisp.dhis.common.DxfNamespaces;
+package org.hisp.dhis.common;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
