+++ dhis-2/dhis-api/src/test/java/org/hisp/dhis/common/DataQueryRequestTest.java
@@ -32,7 +32,6 @@ import static org.hamcrest.Matchers.is;
 
 import org.hisp.dhis.analytics.AggregationType;
 import org.hisp.dhis.analytics.SortOrder;
-import org.hisp.dhis.analytics.UserOrgUnitType;
 import org.joda.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
